refactor: refactor api calls error handling
