+++ src/services/ProductService/ProductService.ts
@@ -1,8 +1,7 @@
 import CtpClient from '../api/BuildClient';
 import { ByProjectKeyRequestBuilder } from '@commercetools/platform-sdk/dist/declarations/src/generated/client/by-project-key-request-builder';
-import { ErrorResponse, ProductProjection, ProductType, TermFacetResult, FacetTerm } from '@commercetools/platform-sdk';
-import { handleErrorResponse } from '../api/handleErrorResponse';
-import { ClientResponse } from '@commercetools/sdk-client-v2';
+import { ProductProjection, ProductType, TermFacetResult, FacetTerm } from '@commercetools/platform-sdk';
+import { getErrorMessage } from '../../utils/helpers/error-handler';
 
 export const FilterParams = {
   category: 'categories.id',
@@ -40,11 +39,11 @@ class ProductService {
       const productTypes = await this.apiRoot.productTypes().get().execute();
       return productTypes.body.results;
     } catch (err) {
-      handleErrorResponse(err as ClientResponse<ErrorResponse> | Error);
+      throw new Error(getErrorMessage(err));
     }
   }
 
-  public async getProduct(ID: string): Promise<ProductProjection | undefined> {
+  public async getProduct(ID: string): Promise<ProductProjection> {
     const productData = await this.apiRoot.productProjections().withId({ ID }).get().execute();
     return productData.body;
   }
@@ -57,7 +56,7 @@ class ProductService {
         .execute();
       return categories.body.results;
     } catch (err) {
-      handleErrorResponse(err as ClientResponse<ErrorResponse> | Error);
+      throw new Error(getErrorMessage(err));
     }
   }
 
@@ -79,7 +78,7 @@ class ProductService {
       );
       return { categories, brands, colors, sizes, prices, products };
     } catch (err) {
-      handleErrorResponse(err as ClientResponse<ErrorResponse> | Error);
+      throw new Error(getErrorMessage(err));
     }
   }
 }
